Refactored vault code, extracted shared code, addressed the code review comments
