+++ internal/pkg/agent/vault/seed.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-//go:build linux || windows
+//go:build !darwin
 
 package vault
 
@@ -14,6 +14,8 @@ import (
 	"os"
 	"path/filepath"
 	"sync"
+
+	"github.com/elastic/elastic-agent/internal/pkg/agent/vault/aesgcm"
 )
 
 const (
@@ -36,8 +38,8 @@ func getSeed(path string) ([]byte, error) {
 	}
 
 	// return fs.ErrNotExists if invalid length of bytes returned
-	if len(b) != int(AES256) {
-		return nil, fmt.Errorf("invalid seed length, expected: %v, got: %v: %w", int(AES256), len(b), fs.ErrNotExist)
+	if len(b) != int(aesgcm.AES256) {
+		return nil, fmt.Errorf("invalid seed length, expected: %v, got: %v: %w", int(aesgcm.AES256), len(b), fs.ErrNotExist)
 	}
 	return b, nil
 }
@@ -59,7 +61,7 @@ func createSeedIfNotExists(path string) ([]byte, error) {
 		return b, nil
 	}
 
-	seed, err := NewKey(AES256)
+	seed, err := aesgcm.NewKey(aesgcm.AES256)
 	if err != nil {
 		return nil, err
 	}
